refactor(parser,core): move getLienHome/getIndexDir/extractRepoId to core
These three utilities resolve where Lien's own state lives (~/.lien,
the per-repo index directory, the repo-id hash) -- none of them depend
on the source code being analyzed, only on decisions Lien itself makes.
parser is meant to be a pure function of the files in front of it;
core is what Lien does with that (storage, config, git). Moves
lien-home.ts, index-dir.ts, and repo-id.ts (with their tests) from
packages/parser/src/utils/ to packages/core/src/utils/, and updates
every internal and cross-package consumer accordingly.

Breaking change to @liendev/parser's public API, accepted without a
deprecation window per this repo's own precedent (pre-1.0, packages
released in lockstep by changesets) -- see the changeset for the full
consumer-sweep rationale.
