+++ packages/core/src/utils/index-dir.ts
