+++ packages/core/src/vectordb/sqlite/sqlite-backend.ts
@@ -2,7 +2,8 @@ import fs from 'fs/promises';
 import path from 'path';
 import type DatabaseType from 'better-sqlite3';
 import type { ChunkMetadata } from '@liendev/parser';
-import { extractRepoId, getLienHome } from '@liendev/parser';
+import { extractRepoId } from '../../utils/repo-id.js';
+import { getLienHome } from '../../utils/lien-home.js';
 import type { SearchResult, VectorDBInterface } from '../types.js';
 import { DatabaseError, wrapError } from '../../errors/index.js';
 import { readVersionFile, writeVersionFile } from '../version.js';
